fix task soltuion
