+++ src/components/TodoFooter/TodoFooter.tsx
@@ -17,7 +17,7 @@ export const TodoFooter: React.FC<Props> = ({
   const activeCount = todos.filter(todo => !todo.completed).length;
   const completedCount = todos.filter(todo => todo.completed).length;
 
-  if (todos.length === 0) {
+  if (!todos.length) {
     return null;
   }
